Update ImportanceSampler._sample_marginal
The previous implementation was summing the conditional probability of
token at position k belonging to the set of specified terms.
This is wrong because we should really be summing over the probabilities
of all sequences containing any of the terms specified. This means that
we should sum over the probability of the current sequence and the fact
that the next token may be one of the specified tokens: p(w_k in W and
w<k).

Whereas previously we were summing over p(w_k in W | w<k)